CompanyController  delete bug fix

+++ src/Core/bizyeriz.Application/Features/Companies/Commands/DeleteCompany/DeleteCompanyQueryHandler.cs
@@ -1,6 +1,7 @@
 ﻿using AutoMapper;
 using bizyeriz.Application.Interfaces.Repositories;
 using bizyeriz.Application.Interfaces.UnitOfWork;
+using bizYeriz.Domain.Entities.CompanyEntities;
 
 namespace bizyeriz.Application.Features.Companies.Commands.DeleteCompany;
 
@@ -18,9 +19,15 @@ public class DeleteCompanyQueryHandler : IRequestHandler<DeleteCompanyQuery, Del
 
     public async Task<DeleteCompanyQueryResponse> Handle(DeleteCompanyQuery request, CancellationToken cancellationToken)
     {
-        var company = await _companyRepository.GetByIdAsync(request.Id, cancellationToken);
+        Company company = await _companyRepository.GetByIdAsync(request.Id, cancellationToken);
+
+        if (company == null)
+        {
+            throw new Exception("Company not found.");
+        }
+
         _companyRepository.Remove(company);
-        _unitOfWork.CommitAsync();
+        await _unitOfWork.CommitAsync();
         DeleteCompanyQueryResponse response = _mapper.Map<DeleteCompanyQueryResponse>(company);
         return response;
     }
